Removed deprecated client and NodeInfo.
