+++ concordium-sdk/src/main/java/com/concordium/sdk/responses/nodeinfo/Node.java
@@ -1,7 +1,6 @@
-package com.concordium.sdk.responses.nodeinfov2;
+package com.concordium.sdk.responses.nodeinfo;
 
 import com.concordium.grpc.v2.NodeInfo;
-import com.concordium.sdk.responses.nodeinfo.ConsensusState;
 import lombok.Builder;
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
